changed intervals to every 5 minutes

+++ db/schema.sql
@@ -2,8 +2,8 @@ DROP TABLE IF EXISTS transactions CASCADE;
 
 CREATE TABLE transactions (
   hash TEXT PRIMARY KEY, 
-  total INT,
-  fees INT,
+  total BIGINT,
+  fees BIGINT,
   inputs JSON,
   outputs JSON
 );
